:recycle: refactor(file): fixtureとしていたテストメソッドの切出し
